consolidate scheduled query code into a single package
